% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.potData}
\alias{as.potData}
\title{Generic Coercion Function}
\usage{
as.potData(object, ...)
}
\arguments{
\item{object}{An object to coerce to the \code{"potData"} class.}

\item{...}{To be passed to methods.}
}
\value{
An object with clas \code{"potData"}.
}
\description{
Coerce an object to the \code{"potData"} class representing
heterogeneous data.
}
