% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potData.R
\name{as.potData.Rendata}
\alias{as.potData.Rendata}
\title{Coerce a \code{Rendata} Object}
\usage{
\method{as.potData}{Rendata}(object, ...)
}
\arguments{
\item{object}{An object having the S3 class \code{"Rendata"} from the
package \strong{Renext}.}

\item{...}{Not used for now. Should in the future allow to set or
reset the names of the blocks using new arguments in
\code{\link{potData}}.}
}
\value{
An object with the S3 class \code{"potData"} from the
package \strong{potomax}.
}
\description{
Coerce a \code{Rendata} object into an object with class
\code{"potData"}.
}
\note{
When \code{object} contains empty OTS blocks a warning is
cast when attempting to find the min of an empty block.
}
\examples{
pdat <- as.potData(Garonne)
pdat1 <- potData(data = Garonne$OTdata$Flow,
                 effDuration = Garonne$OTinfo$effDuration,
                 MAX.data = Garonne$MAXdata$Flow,
                 MAX.effDuration = Garonne$MAXinfo$duration)
all.equal(pdat, pdat1)

## OTS block #2 is empty
set.seed(123)
rd <- rRendata(effDuration = 10,
               OTS.effDuration = rep(60, 2),
               OTS.threshold = rep(80, 2),
               simDate = FALSE,
               distname.y = "GPD", par.y = c(scale = 20, shape = 0.16))
d <- as.potData(rd)
}
