% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotStuff.R
\name{autoplot.confintCheck.poisGP}
\alias{autoplot.confintCheck.poisGP}
\title{Autoplot Method for \code{poisGP} Objects}
\usage{
\method{autoplot}{confintCheck.poisGP}(object, ...)
}
\arguments{
\item{object}{An object with class \code{"confIntCheck.poisGP"}
generated by the \code{confint} method see
\code{\link{confint.poisGP}}.}

\item{...}{Not used yet.}
}
\value{
A ggplot graphics object.
}
\description{
##' Autoplot method for objects representing fitted Poisson-GP models.
}
