% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotStuff.R
\name{autoplot.poisGP}
\alias{autoplot.poisGP}
\title{Autoplot Method for \code{poiGP} Objects}
\usage{
\method{autoplot}{poisGP}(object, which = c("RL", "pp"),
  points = c("H", "p", "none"), a = 0.5, allPoints = FALSE, ...)
}
\arguments{
\item{object}{An object with class \code{"poisGP"}.}

\item{points}{The plotting position system to be used to show the
data attached to the object. When \code{points} is \code{"none"},}

\item{a}{See \code{\link{RP.potData}}.}

\item{...}{}
}
\value{

}
\description{
Autoplot method for objects representing fitted Poisson-Gp models.
}
