% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotStuff.R
\name{autoplot.potData}
\alias{autoplot.potData}
\title{Build ggplot Graphics for a \code{potData} Object}
\usage{
\method{autoplot}{potData}(object,
         type = c("RLplot", "timeplot"),
         aes = FALSE,
         xVar = c("T", "p"),
         group,
         points = c("p", "H"), a = 0.5,
         blockDuration, 
         ...)
}
\arguments{
\item{object}{A \code{potData} object.}

\item{type}{Type of plot wanted, either a return-level plot or a
time plot with the events shown as vertical
segments. \bold{Not implemented yet}.}

\item{aes}{Logical. If \code{TRUE} the colour, fill and shape of
the points are used within the aesthetic function \code{aes} wo
they are registred to appear in legends.}

\item{xVar}{Used when \code{type} is \code{"RLplot"}. The variable
to map the absissa \eqn{x} in the return level plot. When
\code{xVar} is \code{"T"} the plotting positions are computed as
explained in \emph{Renext Computing Details}. For \code{xVar =
"p"} a further step is required to probability of exceedance
\eqn{p} in reference with a given block duration.}

\item{group}{Character with value in \code{"block"} or
\code{"source"}. In the first case the color and the shape of the
points depend on the block. In the second case the color and the
shape depend on the type or source of the block: \code{"OT"},
\code{"MAX"} and \code{"OTS"}. By default the value of
\code{group} is chosen to be \code{"block"} when the number of
block is small enough and \code{"source"} otherwise.}

\item{points}{Type of plotting positions to be used for a RL
plot. See \code{\link{RP.potData}}.}

\item{a}{Parameter to used in \code{ppoints} when \code{points} is
\code{"p"}. See \code{\link{RP.potData}}.}

\item{blockDuration}{A block duration used to find the plotting
positions when \code{xVar} is \code{"p"}.}

\item{...}{Other arguments passed to \code{geom_point}.}
}
\value{
A ggplot layer.
}
\description{
Build ggplot graphics for a \code{potData} object.
}
\note{
The user might have change the colour and the shape of the
points by using \code{scale_colour_manual} and
\code{scale_shape_manual}.
}
\examples{
pdat <- as.potData(Garonne)
autoplot(pdat) + ggtitle("Garonne potData")
g <- autoplot(pdat, aes = TRUE) + ggtitle("Garonne potData")
g
g <- g +  scale_colour_manual(values = c("firebrick", "SpringGreen")) +
       scale_shape_manual(values = c(21, 22))

## use autolayer
ggplot() + autolayer(pdat) + scale_x_log10() + theme_gray()

## Change the label of the historical block
pdat <- potData(data = Garonne$OTdata$Flow, effDuration = 65,
                MAX.data = list("hist" = Garonne$MAXdata$Flow),
                MAX.effDuration = 143)
autoplot(pdat, aes = TRUE)

}
\references{
Yves Deville (2020). \emph{Renext Computing Details}. Technical
Report.
}
\seealso{
\code{\link{potData}}, \code{\link[ggplot2]{scale_manual}}.
}
