% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotStuff.R
\name{autoplot.RL.poisGP}
\alias{autoplot.RL.poisGP}
\title{Autoplot Method for Return Levels}
\usage{
\method{autoplot}{RL.poisGP}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"RL.poisGP"}. It inherits
from \code{data.frame} and has columns c}

\item{...}{Not used yet.}
}
\value{
A ggplot graphic object.
}
\description{
Autoplot method for objects of class \code{"RL.poisGP"}
representing a table of return levels.
}
