% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotStuff.R
\name{autoplot.RL.poisGPList}
\alias{autoplot.RL.poisGPList}
\title{Autoplot method for a list of Return Levels Tables}
\usage{
\method{autoplot}{RL.poisGPList}(object, ...)
}
\arguments{
\item{object}{A list of Return Level tables, with class
\code{"RL.poisGPList"}.}

\item{...}{Not used yet.}
}
\value{
A ggplot graphics object.
}
\description{
Autoplot a list of RL tables, generally using the same data or
using different data but sharing some observations.
}
