% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotStuff.R
\name{autoplot.RLCheck.poisGP}
\alias{autoplot.RLCheck.poisGP}
\title{Autoplot Method for \code{RLCheck.poisGP} Objects}
\usage{
\method{autoplot}{RLCheck.poisGP}(object, ...)
}
\arguments{
\item{object}{An object with class \code{"RLCheck.poisGP"}
generated by the \code{confint} method see
\code{\link{RL.poisGP}}.}

\item{...}{Not used yet.}
}
\value{
A ggplot graphics object.
}
\description{
Autoplot method for objects devoted to the check of the
profile-likelihood confidence intervals on the return levels for
Poisson-GP models.
}
\details{
The object is a list containing two data frames named
\code{RL} and \code{negLogLikC}. The \code{RL} data frame contains
the confidence intervals in a long format suitable for plotting;
it will be used to draw horizontal and vertical lines showing the
confidence limits. The \code{negLogLikC} data frame contains
evaluations of the profile log-likelihood for a grid of value of
each return period; it will be used to display a curve for each
parameter.

The check is similar to that used for the confidence intervals as
implemented in \code{\link{autoplot.confintCheck.poisGP}}. Each
return level is here used as a parameter.
}
\seealso{
\code{\link{RL.poisGP}} and the example with \code{check = TRUE}
in the \bold{Examples} section therein.
\code{\link{autoplot.confintCheck.poisGP}} is a similar method to
check the confidence intervals.
}
