% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisGP.R
\name{coef.poisGP}
\alias{coef.poisGP}
\title{Extract Model Coefficients for a Poisson-GP Object}
\usage{
\method{coef}{poisGP}(object, type = c("poisGP", "PP"), ...)
}
\arguments{
\item{object}{A model with class \code{"poisGP"}.}

\item{type}{The parameterisation to use.}

\item{...}{Not used yet.}
}
\value{
A numeric vector of three coefficients.
}
\description{
Extract model coefficients for a Poisson-GP object.
}
