% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisGP.R
\name{coefIni.poisGP}
\alias{coefIni.poisGP}
\title{Initial Estimates for a \code{poisGP} Model}
\usage{
\method{coefIni}{poisGP}(object, ...)
}
\arguments{
\item{object}{A \code{poisGP} object the parameters of which need
to be estimated.}

\item{...}{Not used yet.}
}
\value{
A vector of parameters for the \code{poisGP} object.
}
\description{
Compute initial estimates for the parameters of a \code{poisGP}
model.
}
\details{
The estimates are obtained as a weighted mean of up to three cheap
estimates corresponding to the different parts of the data: main
sample, MAX blocks and OTS blocks. In each cases, it is assumed
that the shape is zero. For the main sample the estimates of the
rate \code{lambda} and \code{sigma} are simply the exceedance rate
and the mean of the excesses. For MAX and OTS blocks, the
estimates are based on regression approximations as used in
\strong{Renext} package.
}
