% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.poisGP}
\alias{confint.poisGP}
\title{Confidence Intervals for a \code{TVGEV} Object}
\usage{
\method{confint}{poisGP}(object, parm = NULL, level = 0.95,
  method = c("delta", "proflik"), trace = 1L, round = TRUE,
  out = c("array", "data.frame"), ...)
}
\arguments{
\item{object}{An object with class \code{"TVGEV"}.}

\item{parm}{Parameter name. NOT USED YET.}

\item{level}{Confidence level.}

\item{method}{\code{"delta"}, \code{"proflik"},
\code{"boot"}.}

\item{trace}{Integer level of verbosity.}

\item{round}{Logical. If \code{TRUE} the confidence limits will be
rounded to a small number of digits. This number is chosen using the
smallest of the standard deviations for the estimated parameters.}

\item{out}{Character giving the class of the output. By default a
three-dimensional array with dimensions: \emph{parameter},
\emph{lower/upper} limit, and \emph{level}. If \code{level} has length
\code{1}, using \code{drop} on the output will remove the third dimension
and produce a matrix. The \code{"data.frame"} gives the same results
in \code{'long'} format.}

\item{...}{Arguments to be passed to the \code{profLik} or to the
\code{bs} method. For instance, they can be used to chose the type
of bootstrap or the number of bootstrap replications if
\code{method} is \code{"boot"}.}
}
\value{
An array or a data frame with the lower and upper bounds
\code{"L"} and \code{"U"} of the confidence intervals.
}
\description{
Confidence intervals for a \code{TVGEV} object
}
\note{
For the bootstrap method(s), the time required depends on
whether \code{object} embeds a bootstrap distribution or not. When
no bootstrap distribution is found, it has to be computed using
the \code{bs} method; the formal arguments that are given in
\code{\dots} and intended to be used by \code{bs} will then be
ignored with a warning.
}
\examples{


}
