% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{distLines}
\alias{distLines}
\title{Distance Between Two Lines}
\usage{
distLines(x1, x2)
}
\arguments{
\item{x1}{A numeric vector with positive norm.}

\item{x2}{A numeric vector with positive norm and with the same
length as \code{x1}.}
}
\value{
The distance.
}
\description{
Distance between two lines.
}
\details{
This function is used to build diagnostics during the
construction of profile-likelihood confidence intervals. So the
function is not intended to be used by itself.

The vectors \code{x1} and \code{x2} are scaled to have an
unit Euclidean norm and to have the same direction. The distance
between the lines is simply the distance between the unit vectors
on the unit sphere.
}
\note{
When a constrained optimisation is used for the
determination of the lower or upper bound of the interval, we have
to check that the gradient of the objective and that of the
constraint are colinear, which is performed with this function.
}
