% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFuns.R
\docType{data}
\name{Excd}
\alias{Excd}
\title{List of distibutions}
\format{
An object of class \code{list} of length 2.
}
\usage{
Excd
}
\description{
The definitions could easily be put into a loop.
}
\details{
o The first argument is the classical argument of 'd', 'p', q' and
'r' functions. See e.g., \code{link[stats]{Exponential}} distribution.

o The second argument is named \code{theta} and all the
parameters in the right order.
}
\section{Caution}{
 The position of the parameters in the vector
    \code{theta} may matter for some uses for which \code{theta}
    can be unnamed. The first parameter is always assumed to be a
    scale parameter.
}

\keyword{datasets}
