% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEPoisGP.R
\name{lambdaHat}
\alias{lambdaHat}
\title{Compute the ML Estimate of the Rate}
\usage{
lambdaHat(thetaGP, object, log = TRUE, deriv = TRUE)
}
\arguments{
\item{thetaGP}{Parameter vector (GP part only).}

\item{object}{The \code{poisGP} object corresponding to the
estimation.}

\item{log}{Logical. If \code{TRUE} the function returns
\eqn{\log \hat{lambda}}{log(lambdaHat)} possibly with its gradient, else it
returns \eqn{\hat{\lambda}}{lambdaHat}.}

\item{deriv}{Logical. Should the gradient be computed and be
returned as the \code{"gradient"} attribute of the result?}
}
\value{
The value \eqn{\hat{\lambda}}{lambdaHat} of the estimated
rate with an attribute named \code{"r"} corresponding to the
number of observations. When \code{deriv} is \code{TRUE} the
result has an attribute named \code{"gradient"} the value of which
is a row matrix with two columns.
}
\description{
Compute the ML estimate of the rate of the Poisson process given
the GPD parameters.
}
