% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisGP.R
\name{logLik.poisGP}
\alias{logLik.poisGP}
\title{Log-Likelihood of a Poisson-GP Object}
\usage{
\method{logLik}{poisGP}(object, type = c("poisGP", "PP"), ...)
}
\arguments{
\item{object}{An object with class \code{"poisGP"}.}

\item{type}{An experimental argument to make the log-likelihood
comparable to that of other packages/functions. See \strong{Details}.}

\item{...}{Not used yet.}
}
\value{
Value of the log-likelihood.
}
\description{
Log-likelihood of a Poisson-GP model object.
}
\details{
Although models for block maxima and \eqn{r}-largest models can be
considered as special cases of a Poisson-GP model, the
log-likelihood as computed in classical Extreme-Value packages can
differ by a constant related to the observation scheme. The
argument \code{type} aims at finding the same results in the two
cases where the results can be obtained with classical packages,
namely: \emph{"POT" fits with Poisson-GP representation} and
\emph{block maxima or r-largest}. In the first case, this is
hopefully attained by choosing \code{type} to be \code{"poisGP"}
and in the second case by choosing \code{type} to be \code{"PP"}.
}
\section{Caution}{
 Comparing log-likelihoods or related indicators
such AIC, BIC across R packages can be irrelevant due to the use
of different constants.
}

