% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEPoisGP.R
\name{MLE.poisGP}
\alias{MLE.poisGP}
\title{Maximum-Likelihood Estimation of a Poisson-GP Model}
\usage{
\method{MLE}{poisGP}(object = NULL,
    parIni = NULL,
    estim = c("optim", "nloptr", "eval", "none"),
    coefLower, ##  = c("scale" = 0.0, "shape" = -0.99),
    coefUpper, ## = c("scale" = Inf, "shape" = 2.0),
    parTrack =  FALSE,
    scale = FALSE,
    trace = 0)
}
\arguments{
\item{object}{A \code{poisGP} object that needs to be estimated.}

\item{parIni}{Initial values for the parameter vector.}

\item{estim}{Type or method chosen for the estimation.}

\item{coefLower, coefUpper}{Lower and Upper bounds for the
parameters.}

\item{parTrack}{Not used yet.}

\item{scale}{Logical. If \code{TRUE} the data used in the
optimisation are scaled , see \code{\link{threshData}}. NOT
IMPLEMENTED YET.}

\item{trace}{Integer Level of verbosity.}
}
\value{
A list with the results of the likelihood
maximisation. The content of the list depends on the method as
given by \code{estim}, yet it should contain an element
\code{logLik} giving the maximised log-likelihood.
}
\description{
Maximum-Likelihood Estimation of a Poisson-GP model using heterogeneous
data.
}
\details{
The estimation proceeds by minimising a concentrated (or profile)
negative log-likelihood which depends on the two GPD parameters,
but not on the Poisson rate. However the standard (non-profile)
negative log-likelihood function is built and returned because it
will be used to derive profile-likelihood inference results.
}
\author{
Yves Deville
}
