% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{MLE}
\alias{MLE}
\title{Maximum-Likelihood Estimation}
\usage{
MLE(object, ...)
}
\arguments{
\item{object}{An object defining a statistical model, usually
incomplete because the estimation has not yet been performed.}

\item{...}{Arguments passed to methods.}
}
\value{
A list with the results of the estimation. This include an
\code{estimate} element, and usually a \code{cov} element. These
should be extracted by the \code{coef} and \code{vcov} methods.
}
\description{
\code{MLE} is a generic function performing Maximum-Likelihood
estimation of a (usually incomplete) object.
}
\seealso{
\code{\link[stats]{coef}} and \code{\link[stats]{vcov}}.
}
