% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEPoisGP.R
\name{negLogLikFunC}
\alias{negLogLikFunC}
\title{Concentrated or Profile Negative Log-Likelihood Function
for a \code{poisGP} Object}
\usage{
negLogLikFunC(thetaGP, object, deriv = TRUE)
}
\arguments{
\item{thetaGP}{Parameter vector (GP part only).}

\item{object}{The \code{poisGP} object corresponding to the
estimation.}

\item{deriv}{Logical. Should the gradient be computed and be
returned as the \code{"gradient"} attribute of the result?}
}
\value{
The value of the negative log-likelihood. When
\code{deriv} is \code{TRUE} the result has an attribute named
\code{"gradient"} the value of which is a row matrix with two
columns.
}
\description{
Concentrated (or profile) negative log-likelihood function for a
\code{poisGP} object.
}
