% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potData.R
\name{potData}
\alias{potData}
\title{Structured Data Used to Define a Poisson-GP Model}
\usage{
potData(data = NULL, effDuration = NULL,
        MAX.data = NULL, MAX.effDuration = NULL,
        OTS.data = NULL, OTS.threshold = NULL, OTS.effDuration = NULL)
}
\arguments{
\item{data}{A numeric vector containing the observations for the
main sample. If \code{NULL}, the main sample is assumed to be
absent.}

\item{effDuration}{Duration of the main sample.}

\item{MAX.data}{A numeric vector or list of numeric vectors
containing the observations for the \code{MAX} blocks.}

\item{MAX.effDuration}{A numeric vector containing the durations
for the MAX blocks.}

\item{OTS.data}{A numeric vector or list of numeric vectors
containing the observations for the \code{OTS} blocks.}

\item{OTS.threshold}{A numeric vector containing the thresholds
for the OTS blocks.}

\item{OTS.effDuration}{A numeric vector containing the durations for
the OTS blocks.}
}
\value{
An object with class \code{"potDATA"}, essentially a list
with the elements given on input checked and suitably named.
}
\description{
Structured data to define a Poisson-GP model.
}
\details{
As opposed to what is done in \strong{Renext}, no 'main' threshold
is used here. The data is understood as \emph{before} being
censured using any 'main' threshold. So the MAX or OTS blocks can
later be used with any 'main' threshold, even if they contain
observations and thresholds that are smaller than the main
threshold.
}
\examples{

potData(data = Garonne$OTdata$Flow,
        effDuration = 65,
        MAX.data = Garonne$MAXdata$Flow,
        MAX.effDuration = 143)

}
\seealso{
\code{\link{autolayer.potData}} and
\code{\link{autolayer.potData}}.
}
