% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PP2poisGP.R
\name{PP2poisGP}
\alias{PP2poisGP}
\title{Transform Point-Process Parameters into Poisson-GP
Parameters}
\usage{
PP2PoisGP(locStar = 0.0, scaleStar = 1.0, shapeStar = 0.0,
          threshold,
          w = 1.0, deriv = FALSE)
}
\arguments{
\item{locStar, scaleStar, shapeStar}{Numeric vectors containing the
GEV location, scale and shape parameters.}

\item{threshold}{Numeric vector containing the thresholds of the
Poisson-GP model, i.e. the location of the Generalised Pareto
Distribution. \emph{The threshold must be an interior point of the
support of the corresponding GEV distribution}.}

\item{w}{The block duration. Its physical dimension is time and
the product \eqn{\lambda \times w}{\lambda * w} is dimensionless.}

\item{deriv}{Logical. If \code{TRUE} the derivative (Jacobian) of
the transformation is computed and returned as an attribute named
\code{"gradient"} of the attribute.}
}
\value{
A matrix with three columns representing the Poisson-GP
parameters \code{lambda}, code{scale} and \code{shape}.
}
\description{
Transform Point Process (PP) parameters into Poisson-GPparameters.
}
\details{
In the POT framework the three parameters are the rate of
the Poisson process in time: \eqn{\lambda}, and the two GP
parameters: \code{scale} \eqn{\sigma} and \code{shape}
\eqn{\xi}. The vector \code{threshold} contains the fixed
threshold(s) and \code{w} the fixed block duration.
}
\note{
This function is essentially a re-implementation in C of
the function \code{\link[Renext]{gev2Ren}} of \bold{Renext}.
}
\references{
Yves Deville (2020). \emph{Renext Computing Details}. Technical
Report.
}
\author{
Yves Deville
}
