% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisGP.R
\name{print.poisGP}
\alias{print.poisGP}
\title{Print Method for the Class \code{"poisGP"}}
\usage{
\method{print}{poisGP}(x, data = c("none", "raw", "fit"), ...)
}
\arguments{
\item{x}{An object with class \code{"poisGP"}.}

\item{data}{Character. The value \code{"raw"} will lead to print
some information about the raw data as provided to the creator
\code{poisGP}. The value \code{"fit"} will lead to print some
information about the data used for the fit i.e. the excesses over
the threshold.}

\item{...}{Passed to \code{summary}.}
}
\value{
Nothing.
}
\description{
Print an object with class \code{"poisGP"}.
}
\details{
When \code{data} is \code{"fit"} some \code{"OTS"} blocks can be
mentioned while no such blocks exist in the raw data, only MAX
data. This is because as soon as the main threshold exceeds one
observation in a MAX block this block must be turned into an OTS
block with its threshold set to the main threshold.
}
