% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profLik.R
\name{profLik}
\alias{profLik}
\title{Profile-Likelihood Inference Method}
\usage{
profLik(object, fun, ...)
}
\arguments{
\item{object}{An object representing a fitted parametric model.}

\item{fun}{A numeric function of the vector of parameters of the
model given in \code{object}.}

\item{...}{Further arguments for methods.}
}
\value{
The result, typically a numeric array containing
confidence bounds.
}
\description{
Profile-likelihood inference method. This method finds the bounds
of a confidence interval for the given function of the parameter
vector.
}
\details{
The requirement for the method to work is that \code{object} has
suitable "slots", i.e. is a list with suitable elements. These
are: the number \code{p} of parameters, the vector \code{parNames}
of the parameter names, the vector \code{estimate} of ML
estimates, the closure \code{negLogLikFun} computing the negative
log-likelihood function and the value \code{negLogLik} of the
minimised negative log-likelihood. The result returned by
\code{object$negLogLikFun} should have a \code{"gradient"}
attribute to allow the use of derivatives.
}
\author{
Yves Deville
}
