% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLpoisGP.R
\name{RL.poisGP}
\alias{RL.poisGP}
\title{Return Levels and Confidence Intervals for a Poisson-GP Model}
\usage{
\method{RL}{poisGP}(
  object,
  period = NULL,
  level = 0.7,
  confintMethod = c("proflik", "delta", "none"),
  out = c("data.frame", "array"),
  trace = 0,
  ...
)
}
\arguments{
\item{object}{An object with class \code{poisGPML} representing
the inference results for a Poisson-GP model.}

\item{period}{A vector of periods for which the return levels will
be computed.}

\item{level}{Level of confidence. Can be a vector.}

\item{confintMethod}{The method used to compute the confidence
intervals.}

\item{out}{The type of outpout wanted, see the \strong{Value}
section.}

\item{trace}{Integer level of verbosity. The default value is
\code{0}, but when \code{confintMethod} is \code{"proflik"}, it is
a good practice to use \code{trace = 1}, so this may change in the
future.}

\item{...}{Not used yet.}
}
\value{
An object inheriting either from \code{data.frame} or from
\code{array} containing the results. In the first case, the object
has class \code{"RL.poisGP"}, which allows the use of
\code{autoplot} method without recomputing the results.
}
\description{
Compute return levels along with confidence bounds for a Poisson-GP
model with ML inference results.
}
\details{
The return level curve corresponding to the column or to the
dimension named \code{"Quant"} is obtained by plugging the ML
estimate of the Poisson-GP parameters in the quantile. The
confidence limits can be obtained by profile-likelihood or by the
standard 'delta' method.
}
\seealso{
\code{\link{poisGP}} for an example.
}
\author{
Yves Deville
}
