% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{RL}
\alias{RL}
\title{Return Levels associated with Chosen Periods}
\usage{
RL(object, ...)
}
\arguments{
\item{object}{An object defining \emph{return levels} associated
to some period values.}

\item{...}{Arguments passed to methods.}
}
\value{
An object inheriting from the \code{"data.frame"} class
containing the return levels.
}
\description{
Return levels associated with chosen periods.
}
