% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{RP}
\alias{RP}
\title{Return Periods associated with Chosen Quantiles}
\usage{
RP(object, ...)
}
\arguments{
\item{object}{An object defining \emph{return periods} associated
to some "quantiles" values.}

\item{...}{Arguments passed to methods.}
}
\value{
An object inheriting from the \code{"data.frame"} class
contining the return periods.
}
\description{
Return periods associated with chosen quantiles.
}
