% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potData.R
\name{scale.potData}
\alias{scale.potData}
\title{Scale Method for the Class \code{"potData"}}
\usage{
\method{scale}{potData}(x, center = FALSE, scale = FALSE)
}
\arguments{
\item{x}{A \code{potData} object.}

\item{center}{Either the logical \code{FALSE} or a numeric
value. In the later case, \code{center} will first be substracted
to the data.}

\item{scale}{Either the logical \code{FALSE} or a numeric
value. In the later case, \code{scale} the data witll be divided
by \code{scale} after the substraction of \code{center} if
required.}
}
\value{
An object with class \code{"potData"} with the same
structure a \code{x} but with the data rescaled.
}
\description{
Method used to consistently scale each data slot in an object with
class \code{"potData"}.
}
\note{
The scaling of a \code{potData} object can affect the
behaviour of a \code{poisGP} object fitted form it. Fit results
and profile-likelihood results can differ from those expected.
}
\examples{
L <- as.potData(Garonne)
L1 <- scale(L, scale = 1000)

}
