% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManip.R
\name{threshData}
\alias{threshData}
\title{Select the Observations Over a Threshold within
Heterogeneous Data.}
\usage{
threshData(threshold = NULL, data, scale = FALSE)
}
\arguments{
\item{threshold}{A "main" threshold used to select the
observations in each block.}

\item{data}{A list with elements "OT", "MAX" and "OTS". Each
sublist contains a \code{flag} logical element a \code{data}
vector or list and a numeric \code{duration}.}
}
\value{
A list which is comparable to \code{data} but with the
observations below the threshold removed, and the related
information changed. For instance if \code{threshold} is greater
that some observations in a \code{"MAX"} block, these are
discarded and the number \code{r} is changed accordingly.
}
\description{
Select the observations over the given threshold within
heterogeneous data. The data possibly contain OT, MAX and OTS
blocks. The data can optionally be scaled using a scale that is
attached to the result as an attribute.
}
\details{
When \code{scale} is \code{TRUE}, a suitable scale (a positive
number) is chosen as a power of \code{10} and is used to divide
the exceedances over \code{threshold}. This can in some cases
avoid numerical problems.
}
\examples{
set.seed(123)
myData <-
    checkPoisGPData(data = rexp(50), duration = 50,
                    MAX.effDuration = c(25, 10),
                    MAX.data = list("MAX1" = tail(sort(rexp(20))),
                                    "MAX2" = tail(sort(rexp(12))),
                    OTS.effDuration = c(30, 65, 200),
                    OTS.threshold = c(1.8, 3.2, 5.0))
                    OTS.data = list("OTS1" = 1.8 + rexp(5),
                                    "OTS2" = 3.2 + rexp(3),
                                    "OTS3" = 5.0 + rexp(2)))
myData2 <- threshData(threshold = 4, data = myData, scale = FALSE)

}
\seealso{
\code{link{check}}
}
