% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisGP.R
\name{vcov.poisGP}
\alias{vcov.poisGP}
\title{Variance-Covariance Matrix  for a Poisson-GP Object}
\usage{
\method{vcov}{poisGP}(object, type = c("poisGP", "PP"), ...)
}
\arguments{
\item{object}{A model with class \code{"poisGP"}.}

\item{type}{The parameterisation to use.}

\item{...}{Not used yet.}
}
\value{
A covariance matrix of size three.
}
\description{
Variance-covariance matrix for a Poisson-GP object.
}
