% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotStuff.R
\name{autolayer.potData}
\alias{autolayer.potData}
\title{Draw a ggplot Layer for a \code{potData} Object}
\usage{
\method{autolayer}{potData}(object,
          type = c("RLplot", "timeplot"),
          aes = FALSE,
          xVar = c("T", "p"),
          group,
          points = c("p", "H"), a = 0.5,
          ...)
}
\arguments{
\item{object}{A \code{potData} object.}

\item{type}{See \code{\link{autoplot.potData}}.}

\item{aes}{Logical. If \code{TRUE} the colour, fill and shape of
the points are used within the aesthetic function \code{aes}
wo they are registred to appear in legends.}

\item{xVar}{See \code{\link{autoplot.potData}}.}

\item{group}{See \code{\link{autoplot.potData}}.}

\item{points, a}{See \code{\link{autoplot.potData}}.
' \code{"p"}. See \code{\link{RP.potData}}.}

\item{...}{Other arguments passed to \code{geom_point}.}
}
\value{
A ggplot object.
}
\description{
Draw a ggplot layer for a \code{potData} object.
}
\note{
With \code{aes = TRUE} it can be needed to reset the scales
to avoid the creation of confusing legends. The package
\strong{ggnewscale} can be of some help then.
}
\examples{
fit <- poisGP(data = Garonne, threshold = 2800)
RL <- RL(fit)
g <- autoplot(RL)
g
pdat <- as.potData(Garonne)
g1 <- g + autolayer(pdat, aes = TRUE)
g1

}
\seealso{
\code{\link{potData}}, \code{\link{autoplot.potData}}.
}
