% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotStuff.R
\name{autoplot.confintCheck.poisGP}
\alias{autoplot.confintCheck.poisGP}
\title{Autoplot Method for \code{confintCheck.poisGP} Objects}
\usage{
\method{autoplot}{confintCheck.poisGP}(object, ...)
}
\arguments{
\item{object}{An object with class \code{"confintCheck.poisGP"}
generated by the \code{confint} method see
\code{\link{confint.poisGP}}.}

\item{...}{Not used yet.}
}
\value{
A ggplot graphics object.
}
\description{
Autoplot method for objects devoted to the check of
    the profile-likelihood confidence intervals on the parameters
    for Poisson-GP models.
}
\details{
The object is a list containing two data frames named \code{ci}
and \code{negLogLikC}. The \code{ci} data frame contains the
confidence intervals in a long format suitable for plotting; it
will be used to draw horizontal and vertical lines showing the
confidence limits. The \code{negLogLikC} data frame contains
evaluations of the profile log-likelihood for a grid of value of
each parameter; it will be used to display a curve for each
parameter.

For each parameter, the two verical lines corresponding to a given
confidence level represent the lower and upper bounds of the
confidence interval for the parameter. The curve represent the
negative profiled log-likelihood for the considered parameter. For
each confidence level the two vertical lines should intersect the
curve at points where the value of the profiled negative
log-likelhood corrresponds to the horizontal line for the the
level. The horizontal and vertical lines for the estimate should
intersect at the point where the value of the negative
log-likelhood is minimal. If the later condition is not true then
the convergence of the ML estimation was not reached for some
reason.
}
\seealso{
\code{\link{confint.poisGP}} and the example with
    \code{check = TRUE} in the \bold{Examples} section therein.
}
