% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.poisGP}
\alias{confint.poisGP}
\title{Confidence Intervals for a \code{poisGP} Object}
\usage{
\method{confint}{poisGP}(object,
        parm = NULL,
        level = 0.95,
        type = c("poisGP", "PP"),
        method = c("proflik", "delta"),
        nSigma = 4,
        trace = 0,
        round = TRUE,
        out = c("array", "data.frame"),
        check = FALSE, nCheck = 50,
        ...)
}
\arguments{
\item{object}{An object with class \code{"poisGP"}.}

\item{parm}{Not used yet. Confidence intervals are computed for
each of the three parameters.}

\item{level}{Confidence level(s).}

\item{type}{The type of parameterisation wanted: \emph{Poisson-GP}
or \emph{Point-Process}.}

\item{method}{Character: \code{"delta"} leads to the simplistic
\emph{delta} method and \code{"proflik"} to the
\emph{profile-likelihood}.}

\item{nSigma}{Used only when \code{check} is \code{TRUE}. It
defines the interval around the ML estimate where the profile
log-likelihood will be evaluated in order to build a curve
allowing a check of the results. The value of \code{nSigma}
defines the number of standard deviation for the current parameter
that will be used. If needed, an an asymmetric interval can be
defined by using two numbers e.g. \code{c(3, 5)} if it is expected
that the confidence intervals spread more on the right side of the
estimates than they do on the left side.}

\item{trace}{Integer level of verbosity.}

\item{round}{Logical. If \code{TRUE} the confidence limits will be
rounded to a same small number of digits. This number is chosen
using the smallest of the standard deviations for the estimated
parameters.}

\item{out}{Character giving the class of the output. By default
this is a three-dimensional array with dimensions:
\emph{parameter}, \emph{lower/upper} limit, and \emph{level}. If
\code{level} has length \code{1}, using \code{drop} on the output
will remove the third dimension and produce a matrix. The
\code{"data.frame"} gives the same results in "long
format".}

\item{check}{Logical. Used only when \code{method} is
\code{"proflik"}. If \code{TRUE} the function return results
intended to be used in a graphical check of the confidence limits
and taking the form of a list of two data frames. The first data
frame contains evaluations of the profile-negative log-likelihood
for each of the three parameters in order to draw
profile-likelihood curves. The second one contains the confidence
bounds in "long format".}

\item{nCheck}{Number of evaluations of the profile log-likelihood if
\code{check} is \code{TRUE}.}

\item{...}{Not used yet.}
}
\value{
When \code{check} is \code{FALSE}, an array or a data
frame containing the lower and upper bounds \code{"L"} and
\code{"U"} of the confidence intervals. When \code{check} is
\code{TRUE} a list of two data frames which is given the class
\code{"confintCheck"} is order to use the \code{autoplot} method
that is implemented for this class.
}
\description{
Confidence intervals for the parameters of a \code{poisGP} object.
}
\details{
This method finds confidence intervals for each of the three
parameters of the chosen parameterisation: \code{"poisGP"}
(default) or \code{"PP"}. The recommended (and default) method
uses profile-likelihood as implemented in \code{proflik.default}.
The determination of the intervals relies on a new method based on
constrained optimisation: thus the profiled likelihood is not
computed as such, as opposed to what is usually done. The profiled
likelihood can be computed to check the results by using
\code{check = TRUE}, but no zero finding is used to find the
confidence limits following the classical method. The check is
thus entirely based on the graphics which must be carefully
inspected.
}
\note{
Remind that the \code{"PP"} parameterisation does not depend
on the threshold, as opposed to the \code{"poisGP"}
parameterisation. So \code{type = "PP"} should be used to
investigate \emph{threshold stability} for the full parameter
vector. However the confidence intervals on the two shape
parameter: Poisson-GP \eqn{\xi} and PP \eqn{\xi^\star}{\xi*} are
(or should be) identical.
}
\section{Caution}{
 The determination of the profile-likelihood
intervals can fail, so it is wise to set \code{check = TRUE} and
use the \code{autoplot} method on the returned object. Problems
seem to be more frequently met with \code{type = "PP"}, i.e. when
the Point-Process parameterisation is used.
}

\examples{
## fit from the object Garonne from Renext (class "Rendata")
fit <- poisGP(Garonne, threshold = 2900)

ci <- confint(fit, lev = c(0.70, 0.95), trace = 1)

## Check the results: this is quite time-consuming.
\dontrun{
cic <- confint(fit, lev = c(0.95, 0.70), check = TRUE)
autoplot(cic) + theme_gray()

cicPP <- confint(fit, type = "PP", lev = c(0.95, 0.70), check = TRUE)
autoplot(cicPP) + theme_gray()
}
}
\seealso{
\code{\link{RL.poisGP}} for the computation of the return
levels with confidence intervals,
\code{\link{autoplot.confintCheck.poisGP}} for the graphical check
of the results. The \code{profLik.default} method is used by this
function.
}
