% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEPoisGP.R
\name{MLE.poisGP}
\alias{MLE.poisGP}
\title{Maximum-Likelihood Estimation of a Poisson-GP Model}
\usage{
MLE.poisGP(object = NULL, parIni = NULL, estim = c("optim", "nloptr",
  "none"), coefLower = c(scale = 0, shape = -0.99), coefUpper = c(scale
  = Inf, shape = 2), parTrack = FALSE, trace = 0)
}
\arguments{
\item{object}{A \code{poisGP} object that needs to be estimated.}

\item{parIni}{Initial values for the parameter vector.}

\item{estim}{Type or method chosen for the estimation.}

\item{coefLower, coefUpper}{Lower and Upper bounds for the
parameters.}

\item{parTrack}{Not used yet.}

\item{trace}{Integer Level of verbosity.}
}
\value{
A list with the results of the likelihood
maximisation. The content of the list depends on the method as
given by \code{estim}, yet it should contain an element
\code{logLik} giving the maximised log-likelihood.
}
\description{
Maximum-Likelihood Estimation of a Poisson-GP model using heterogeneous
data.
}
\details{
The estimation proceeds by minimising a concentrated (or profile)
negative log-likelihood which depends on the two GPD parameters
but not on the Poisson rate. However the negative log-likelihood
function is built and returned because it will be used to derive
profile-likelihood inference results.
}
\author{
Yves Deville
}
