% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEPoisGP.R
\name{MLE.poisGP}
\alias{MLE.poisGP}
\title{Maximum-Likelihood Estimation of a Poisson-GP Model}
\usage{
\method{MLE}{poisGP}(object = NULL,
    parIni = NULL,
    estim = c("optim", "nloptr", "eval", "none"),
    coefLower, coefUpper,
    parTrack =  FALSE,
    scale = FALSE,
    trace = 0)
}
\arguments{
\item{object}{A \code{poisGP} object that needs to be estimated.}

\item{parIni}{Initial values for the parameter vector. This
is must be a named vector of length \eqn{2} with elements
names \code{"scale"} and \code{"shape"}.}

\item{estim}{Type or method chosen for the estimation.}

\item{coefLower, coefUpper}{Lower and Upper bounds for the
parameters. The should be numeric vectors with names in
\code{c("lambda", "scale", "shape")}. Only the bounds on the GP
parameters \code{"scale"} and \code{"shape"} can be used during
the estimation and they will only be used when \code{estim} is
\code{"nloptr"}. However the bounds are used in the inference
\code{\link{confint.poisGP}} and \code{\link{RL.poisGP}}.}

\item{parTrack}{Not used yet.}

\item{scale}{Logical. If \code{TRUE} the data used in the
optimisation are scaled , see \code{\link{threshData}}. NOT
IMPLEMENTED YET.}

\item{trace}{Integer Level of verbosity.}
}
\value{
A list with the results of the likelihood
maximisation. The content of the list depends on the method as
given by \code{estim}, yet it should contain an element
\code{logLik} giving the maximised log-likelihood.
}
\description{
Maximum-Likelihood Estimation of a Poisson-GP model using heterogeneous
data.
}
\details{
The estimation proceeds by minimising a concentrated (or profile)
negative log-likelihood which depends on the two GPD parameters,
but not on the Poisson rate. So provided bound on this parameter
will have no effect on the estimation and the estimates can fail
to have its element \code{"lambda"} within the bounds when these
are not \code{0.0} and \code{Inf}. However the standard
(non-profile) negative log-likelihood function is built and
returned because it will be used to derive profile-likelihood
inference results.
}
\author{
Yves Deville
}
