% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisGP.R
\name{poisGP}
\alias{poisGP}
\title{Create a Poisson-GP Model Object}
\usage{
poisGP(data = NULL, threshold, effDuration, MAX.data = NULL,
  MAX.effDuration = NULL, OTS.data = NULL, OTS.threshold = NULL,
  OTS.effDuration = NULL, parIni = NULL, estim = c("optim", "nloptr",
  "none"), coefLower = c(scale = 0, shape = -0.9), coefUpper = c(scale
  = Inf, shape = Inf), trace = 0)
}
\arguments{
\item{data}{A vector containing the observations for the "main" sample.}

\item{threshold}{The \emph{main threshold} as commonly understood
in POT.}

\item{effDuration}{The effective duration of the observation
period corresponding to the \code{data}.}

\item{MAX.data}{A list of numeric vectors corresponding to periods
or \emph{blocks}. Each vector contains the \eqn{r}-largest
observations for the block, where \eqn{r>0} can vary across
blocks. When a numeric vector is passed instead of a list, it is
understood that there is only one MAX block.}

\item{OTS.data}{A list of numeric vectors corresponding to periods
or \emph{blocks}. Each vector contains all observations for the
block which exceeded the corresponding threshold as given in
\code{OTS.threshold}. So the number \eqn{r \eg 0} vary across
blocks. When a numeric vector is passed instead of a list, it is
understood that there is only one OTS block.}

\item{OTS.threshold}{A vector of positive numbers giving the
thresholds of the OTS blocks (in the same order). By construction
all the elements of \code{OTS.data[i]} are larger than
\code{OTS.threshol[i]} for each block index \code{i}.}

\item{parIni}{A named parameter vector. This will be used to set
the values if \code{estim} is \code{"none"} or to provide initial
values in the other cases. When \code{parIni} is \code{NULL} the
parameter vector stored in the object as will contain \code{NA}
when \code{estim} is \code{"none"} or "good" initial values found
by devoted function.}

\item{estim}{Method or XXX}

\item{coefLower, coefUpper}{Named vectors of bounds for the
parameters. The values can be infinite \code{Inf} or \code{-Inf}.
However, note that without bounds on the shape parameter \eqn{xi} the
maximum likelihood is infinite and obtained for \eqn{xi = -1}.}

\item{trace}{Integer level of verbosity.}

\item{MAX.duration}{A vector of positive numbers giving the
durations of the MAX blocks (in the same order).}

\item{OTS.duration}{A vector of positive numbers giving the
durations of the OTS blocks (in the same order).}
}
\value{
A list with among which

\item{data}{

A copy of the data provided in \code{data} and the other formals
\code{MAX.*} or \code{OTS.*}

}
\item{fitData}{

A modified version of \code{data} where the observations that do
not exceed the main threshold \eqn{u} are discarded and each
remaining observation \eqn{y_i} is replaced by the corresponding
excess \eqn{y_i - u}. So only positive observations are found in
the data vectors.

}
}
\description{
Create a Poisson-GP model object, usually by ML estimation.
}
\examples{

fit <- poisGP(data = Garonne$OTdata$Flow, threshold = 2900,
              effDuration = 65,
              MAX.data = Garonne$MAXdata$Flow,
              MAX.effDuration = 143,
              estim = "none")



}
\author{
Yves Deville
}
