% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profLik.R
\name{profLik}
\alias{profLik}
\title{Profile-Likelihood Inference Method}
\usage{
profLik(object, fun, ...)
}
\arguments{
\item{object}{An object representing a fitted parametric model.}

\item{fun}{A numeric function of the vector of parameters of the
model given in \code{object}.}

\item{...}{Further arguments for methods.}
}
\value{
The result, typically a numeric array containing
confidence bounds.
}
\description{
Profile-likelihood inference method. This method finds the bounds
of a confidence interval for the given function of the parameter
vector.
}
\details{
Under suitable conditions such as the smoothness of the function
\eqn{f(\boldsymbol{\theta})}{f(\theta}) given in \code{fun},
\eqn{\eta := \boldsymbol{\theta}}{\eta := f(\theta)} can be
considered as a parameter of the model in a suitable
re-parameterisation of it. So it makes sense to use the
profile-likelihood method to derive confidence intervals on
it. Although different methods can be used for this the
\bold{potomax} package flavours using an optimisation of
\eqn{f(\boldsymbol{\theta})}{f(\theta}) under a constraint of high
log-likelihood \eqn{\ell(\boldsymbol{\theta}) \geq
\ell_{\textrm{max}} - \delta}{l(\theta) >= lMax - \delta} where
\eqn{\delta} is a small positive value depending on the confidence
level.
}
\author{
Yves Deville
}
