% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLpoisGP.R
\name{RL.poisGP}
\alias{RL.poisGP}
\title{Return Levels and Confidence Intervals for Poisson-GP Model}
\usage{
\method{RL}{poisGP}(object, period = NULL, level = 0.7,
  confintMethod = c("delta", "none", "boot", "proflik"),
  out = c("data.frame", "array"), biasCorrect = FALSE, trace = 1L,
  ...)
}
\arguments{
\item{object}{An object with class \code{poisGPML} representing
the inference results for a Poisson-GP model.}

\item{period}{A vector of periods for which the return levels will
be computed.}

\item{level}{Level of confidence.}

\item{confintMethod}{The method used to compute the confidence
intervals.}

\item{out}{The type of outpout wanted, see the \strong{Value}
section.}

\item{biasCorrect}{Logical. Should bias correction be applied in
bootstrap?}

\item{trace}{Integer level of verbosity.}

\item{...}{Not used yet.}
}
\value{
An object inheriting either from \code{data.frame} or from
\code{array} containing the results. In the first case, the object
has class \code{"RL.poisGP"}, which allows the use of
\code{autoplot} method without recomputing the results.
}
\description{
Compute return levels along with confidence bounds for a Poisson-GP
model with ML inference results.
}
\details{
The return level curve corresponding to the column or to the
dimension named \code{"Quant"} is obtained by plugging the ML
estimate of the poisson-GP parameters in the quantile. To a
certain extension, this can be compared to a Bayesian MAP when a
flat prior is used. However this is not exactly the case because
the ML Estimate is invariant under reparameterising, while a flat
prior refers to a specific parameterisation, hence so does the
corresponding MAP. The profile-likelihood is obtained by using a
specific algorithm based on constrained optimisation, and not on
the usual method of re-parameterisartion with quantiles.
}
\author{
Yves Deville
}
