% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{RL}
\alias{RL}
\title{Return Levels associated with Chosen Periods}
\usage{
RL(object, ...)
}
\arguments{
\item{object}{An object that can be used to define \emph{return
levels} associated to some period values.}

\item{...}{Arguments passed to methods.}
}
\value{
An object inheriting from the \code{"data.frame"} class
containing the return levels.
}
\description{
\code{RL} is a generic function which computes the return levels
associated with chosen periods using a suitable model or data set.
}
